Allow command-line arguments to be passed in a file
This take the gcc format of @<filename>. If <filename> does not exist or cannot be opened, the argument is accepted as-is. The file (or ISPC_ARGS) is itself allowed to contain another @<filename> argument.
The content of the file is split into arguments in essentially the same way as was used for ISPC_ARGS. It does not (yet) support any method of escapong or quoting to allow white space within an argument.
This also now polices that meximum number of arguments allowed rather than corrupting memory!