Implement GCC-style makefile dependency generation
This commit adds handling of commandline options -M and -MF in a way inspired by what GCC does.
-M generates a makefile-compatible target rule for the current output object (or the canonical a.out, if nothing is specified), and outputs it to stdout.
-MF <filename> puts the output in <filename> instead.
The functionality of -MMM remains unchanged, apart from the fact that a subsequent -MF may alter the filename.
In order to achieve this, a new bitflag type Module::OutputFlags is introduced. It replaces the generatePIC boolean and is added as a new argument in several places.