Fix synchronisation issue in run_tests.py.
The pattern with checking Queue.empty() and doing Queue.get() without critical
section leads to dead locks. So, options are: use critical section or different
mechanism to pull the new job. This solution uses "STOP" markers in the Queue.
So, no check for empty queue is needed. The same queue is also used to indicate
that all tests are handled, as it's JoinableQueue. Hence getting rid of extra
Queue for signaling completion.

Also queue for signalling errors is removed. All exception will be reported as
Runtime failures and reported in the log.