Update bench workflow to trigger calcite diff only when it should
Since we changed the condition to use the `pull_request_target` event, the parent commit of the build is not necessarily available (since it uses the target branch). So the trigger (which should trigger a diff event and notifications) fails since there’s no data for the parent commit.
For now, set the condition back to `pull_request` so that we can enable it later once we confirm what type of flow we want for notifications.