ispcrt: Add TaskQueue::submit() method
TaskQueue::sync() method submits command queue for execution
and immediately blocks waiting for the execution to complete.

If app wants to submit lists from different queues or different
devices, it needs to use threading or other tricks.
This commit adds TaskQueue::submit() method which submits the queue
for execution and immediatelly returns.
The behavior of sync() is not changed - it will call submit() if needed.